FIX: fixed typing issue with downsample scale. When a task is loaded, integer types are loaded as np.int64. Having downsample=1 and downsampling.enabled=False caused the PMA loop to return a float64 array instead of the desired float32 array. I updated the PMA class to deal with this, but I should really be updating the loader instead.

+++ src/pyxalign/alignment/projection_matching.py
@@ -70,7 +70,7 @@ class ProjectionMatchingAligner(Aligner):
             center_of_rotation=self.projections.center_of_rotation,
         )
         if self.options.downsample.enabled:
-            self.scale = self.options.downsample.scale
+            self.scale = (self.options.downsample.scale).astype(int)
         else:
             self.scale = 1
